+++ scripts/systems/pf2e.js
@@ -1,6 +1,8 @@
 import { Stealthy } from '../stealthy.js';
 import Engine from '../engine.js';
 
+async function wait(ms) { return new Promise(resolve => { setTimeout(resolve, ms); }); }
+
 export class EnginePF2e extends Engine {
 
   constructor() {
@@ -8,33 +10,34 @@ export class EnginePF2e extends Engine {
 
     // There is probably a better practice for figuring out skill checks in PF2E, but this "works"
     Hooks.on('createChatMessage', async (message, options, id) => {
-      if (message.flavor.includes('>Skill Check: Stealth<')) {
+      // Stealthy.log("createChatMessage", message);
+      if (['>Skill Check: Stealth<', '>Initiative: Stealth<', '>(Stealth Check)<'].some(t => message.flavor.includes(t))) {
         await this.rollStealth(message, options, id);
       }
-      else if (message.flavor.includes('>Perception Check<')) {
+      else if (['>Perception Check<', '>(Perception Check)<'].some(t => message.flavor.includes(t))) {
         await this.rollPerception(message, options, id);
       }
     });
   }
 
   findHiddenEffect(actor) {
-    return actor.getFlag(Stealthy.MODULE_ID, 'hidden');
+    return actor.getCondition('hidden');
   }
 
   findSpotEffect(actor) {
-    return actor.getFlag(Stealthy.MODULE_ID, 'spot');
+    return null;
   }
 
   canSpotTarget(visionSource, hiddenEffect, target) {
-    const source = visionSource.object?.actor;
-    const stealth = hiddenEffect ?? (10 + target.actor.system.skills.ste.value);
-    const spotEffect = this.findSpotEffect(source);
-    const perception = spotEffect ?? (10 + target.actor.system.attributes.perception.value);
+    // const source = visionSource.object?.actor;
+    // const stealth = hiddenEffect ?? (10 + target.actor.system.skills.ste.value);
+    // const spotEffect = this.findSpotEffect(source);
+    // const perception = spotEffect ?? (10 + target.actor.system.attributes.perception.value);
 
-    if (perception <= stealth) {
-      Stealthy.log(`${visionSource.object.name}'s ${perception} can't see ${target.name}'s ${stealth}`);
-      return false;
-    }
+    // if (perception <= stealth) {
+    //   Stealthy.log(`${visionSource.object.name}'s ${perception} can't see ${target.name}'s ${stealth}`);
+    //   return false;
+    // }
     return true;
   }
 
@@ -47,19 +50,25 @@ export class EnginePF2e extends Engine {
   }
 
   async updateOrCreateEffect({ label, actor, flag, source, makeEffect }) {
-    for (const [key, value] of Object.entries(flag)) {
-      await actor.setFlag(Stealthy.MODULE_ID, key, value);
-      break;
+    const lowerLabel = label.toLowerCase();
+    if (!actor.hasCondition(lowerLabel)) {
+      await actor.toggleCondition(lowerLabel);
     }
+    const condition = actor.getCondition(lowerLabel);
+    let update = duplicate(condition.toObject(false));
+    update.flags.stealthy = flag;
+    await actor.updateEmbeddedDocuments('Item', [update]);
   }
 
   getHiddenFlagAndValue(actor, effect) {
-    const value = effect ?? (10 + actor.system.skills.ste.value);
-    return { value };
+    const value = effect.flags.stealthy?.hidden ?? (10 + actor.system.skills.ste.value);
+    return { flag: { hidden: value }, value };
   }
 
   async setHiddenValue(actor, effect, flag, value) {
-    await actor.setFlag(Stealthy.MODULE_ID, 'hidden', value);
+    flag.hidden = value;
+    effect.flags.stealthy = flag;
+    await actor.updateEmbeddedDocuments('Item', [effect]);
     stealthy.socket.executeForEveryone('RefreshPerception');
   }
 
@@ -75,20 +84,30 @@ export class EnginePF2e extends Engine {
 
   async rollPerception(message, options, id) {
     Stealthy.log('rollPerception', { message, options, id });
+    // const check = Number(message.content);
 
-    const token = canvas.tokens.get(message.speaker.token);
-    const actor = token.actor;
-    await this.updateOrCreateSpotEffect(actor, { spot: Number(message.content) });
+    // const token = canvas.tokens.get(message.speaker.token);
+    // const actor = token.actor;
+    // await this.updateOrCreateSpotEffect(actor, { spot: check });
 
     super.rollPerception();
   }
 
   async rollStealth(message, options, id) {
     Stealthy.log('rollStealth', { message, options, id });
+    const check = Number(message.content);
 
     const token = canvas.tokens.get(message.speaker.token);
     const actor = token.actor;
-    await this.updateOrCreateHiddenEffect(actor, { hidden: Number(message.content) });
+    await this.updateOrCreateHiddenEffect(actor, { hidden: check });
+    // const label = 'hidden';
+    // if (!actor.hasCondition(label)) {
+    //   await actor.toggleCondition(label);
+    // }
+    // const hidden = actor.getCondition(label);
+    // let update = duplicate(hidden.toObject(false));
+    // update.flags.stealthy = check;
+    // await actor.updateEmbeddedDocuments('Item', [update]);
 
     super.rollStealth();
   }